bio_info/apps/ukb-api/pyproject.toml
[project]
name = "ukb-api"
version = "0.1.0"
description = "UK Biobank data service with REST API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "dx-client",
    "pandas>=2.2.0",
    "pydantic-settings>=2.7.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "duckdb==1.4.4",
    "sqlglot>=30.4.2",
]

[project.scripts]
ukb-api = "ukb_api.cli:main"


[tool.fastapi]
entrypoint = "ukb_api.main:app"
host = "0.0.0.0"
port = 8000

[tool.fastapi.serve]
host = "0.0.0.0"
port = 8000

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"
